requirements.txt
aio-pika==9.4.1
aiormq==6.8.1
altgraph==0.17.4
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
multidict==6.7.0
outcome==1.3.0.post0
packaging==25.0
pamqp==3.3.0
pefile==2023.2.7
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
requests==2.32.5
selenium==4.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tk==0.1.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wheel==0.45.1
wsproto==1.3.1
yarl==1.22.0
